chore: clippy cleanup (20 lint categories)
Cleared 20 distinct clippy lint categories across the workspace,
accumulated over many small autonomous-loop iterations. All instances
of each lint eliminated; no `#[allow]` suppressions added.

Lints cleared (counts as observed at time of fix):
- needless_return (41 sites): drop trailing `return X;` in match arms.
- needless_borrows_for_generic_args (4): `pow(&[i])` -> `pow([i])`.
- manual_repeat_n (7): `repeat(x).take(n)` -> `repeat_n(x, n)`.
- clone_on_copy (15 in scope): drop `.clone()` on Copy types in
  examples/kzg, examples/pedersen_eq, benches/graph_execution.
- manual_div_ceil (3): `(a + b - 1) / b` -> `a.div_ceil(b)`.
- manual_is_multiple_of (8): `x % n == 0` -> `x.is_multiple_of(n)`.
- to_string_in_format_args (2): drop `.to_string()` in format! args.
- redundant_guard (2): `Index(i) if i == 0` -> `Index(0)`.
- expect_fun_call (8): `.expect(format!(...).as_str())` ->
  `.unwrap_or_else(|| panic!(...))` (avoids eager format).
- let_and_return (8): inline trailing `let x = E; x` returns.
- len_zero (2): `.len() > 0` -> `!.is_empty()`.
- match_like_matches_macro (3): match-bool -> `matches!`.
- unnecessary_cast (9): drop `as T` self-casts.
- single_match (3): collapse to `if let`.
- map_clone (8): `.map(|x| x.clone())` -> `.cloned()`.
- implied_bounds_in_impls (2): drop Iterator+DoubleEndedIterator
  redundancy.
- unnecessary_unwrap (1): merge `is_ok() + unwrap()` into `if let Ok`.
- collapsible_if (13): nested `if let .. { if .. }` -> let-chains.
- redundant_closure (15): `.map(|b| ATyp::Base(b))` -> `.map(ATyp::Base)`.
- into_iter_on_ref (15): `.into_iter()` -> `.iter()`/`.iter_mut()` on
  borrows of Vec.
- derivable_impls (1): manual `impl Default for Distribution` ->
  `#[derive(Default)]` with `#[default]` on `Nonuniform`.
- needless_lifetimes (35): elide inferable explicit lifetimes.
- useless_conversion (32): drop redundant `.into()`/`.into_iter()`.
- unnecessary_mut_passed (75): pass `&ctx` instead of `&mut ctx` where
  the callee signature takes `&Self::Context`.

Build, fmt, and full test suite (980+ tests) all green.

+++ graph/src/scheduler/asymptotic_cost.rs
@@ -188,13 +188,13 @@ impl<C: ArkConfig> CostModel<C, Ref> for AsymptoticCost<C> {
             }
             Op::Ifft(op) => {
                 let n = op.typ().size() as f64;
-                cost += self.cost(op, nthreads).0
-                    + (n * (n as f64).log2() * Self::SCALAR_MUL / nthreads as f64)
+                cost +=
+                    self.cost(op, nthreads).0 + (n * n.log2() * Self::SCALAR_MUL / nthreads as f64)
             }
             Op::Fft(op) => {
                 let n = op.typ().size() as f64;
-                cost += self.cost(op, nthreads).0
-                    + (n * (n as f64).log2() * Self::SCALAR_MUL / nthreads as f64)
+                cost +=
+                    self.cost(op, nthreads).0 + (n * n.log2() * Self::SCALAR_MUL / nthreads as f64)
             }
             Op::Check(op) => cost += self.cost(op, nthreads).0,
             Op::Poly(_op) => cost += 1.0,
